speed up the building step of GannIndex

+++ gann.go
@@ -31,8 +31,8 @@ import (
 
 type GannIndex interface {
 	Build() error // build search trees.
-	GetANNbyItemID(id int64, num int, bucketScale float64) (ann []int64, err error)
-	GetANNbyVector(v []float32, num int, bucketScale float64) (ann []int64, err error)
+	GetANNbyItemID(id int64, num int, bucketScale float32) (ann []int64, err error)
+	GetANNbyVector(v []float32, num int, bucketScale float32) (ann []int64, err error)
 }
 
 // GetIndex ... get index (composed of trees, nodes, etc.)
